refactoring into module directory + coverage tests
