+++ src/controllers/module/moduleController.ts
@@ -1,7 +1,7 @@
 import {Request, Response, Router} from 'express'
-import {ModuleFactory} from '../learning-catalogue/model/factory/moduleFactory'
+import {ModuleFactory} from '../../learning-catalogue/model/factory/moduleFactory'
 import * as log4js from 'log4js'
-import {LearningCatalogue} from '../learning-catalogue'
+import {LearningCatalogue} from '../../learning-catalogue'
 
 const logger = log4js.getLogger('controllers/moduleController')
