+++ src/controllers/module/youtubeModuleController.ts
@@ -28,6 +28,7 @@ export class YoutubeModuleController {
 		this.setRouterPaths()
 	}
 
+	/* istanbul ignore next */
 	private setRouterPaths() {
 		this.router.param('courseId', async (req, res, next, courseId) => {
 			const course = await this.learningCatalogue.getCourse(courseId)
