+++ test/unit/controllers/module/youtubeModuleControllerTest.ts
@@ -1,17 +1,17 @@
 import * as chai from 'chai'
 import * as sinonChai from 'sinon-chai'
-import {YoutubeModuleController} from '../../../src/controllers/module/youtubeModuleController'
-import {ModuleFactory} from '../../../src/learning-catalogue/model/factory/moduleFactory'
-import {LearningCatalogue} from '../../../src/learning-catalogue'
+import {YoutubeModuleController} from '../../../../src/controllers/module/youtubeModuleController'
+import {ModuleFactory} from '../../../../src/learning-catalogue/model/factory/moduleFactory'
+import {LearningCatalogue} from '../../../../src/learning-catalogue'
 import {mockReq, mockRes} from 'sinon-express-mock'
-import {ContentRequest} from '../../../src/extended'
+import {ContentRequest} from '../../../../src/extended'
 import * as sinon from 'sinon'
-import {Course} from '../../../src/learning-catalogue/model/course'
+import {Course} from '../../../../src/learning-catalogue/model/course'
 import {Request, Response} from 'express'
 import {expect} from 'chai'
-import {Module} from '../../../src/learning-catalogue/model/module'
+import {Module} from '../../../../src/learning-catalogue/model/module'
 import {YoutubeService} from 'lib/youtubeService'
-import {Validator} from '../../../src/learning-catalogue/validator/validator'
+import {Validator} from '../../../../src/learning-catalogue/validator/validator'
 
 chai.use(sinonChai)
